+++ libheif/linux_build_libs.py
@@ -15,7 +15,7 @@ PH_LIGHT_VERSION = sys.maxsize <= 2**32 or getenv("PH_LIGHT_ACTION", "0") != "0"
 LIBX265_URL = "https://bitbucket.org/multicoreware/x265_git/get/0b75c44c10e605fe9e9ebed58f04a46271131827.tar.gz"
 LIBAOM_URL = "https://aomedia.googlesource.com/aom/+archive/v3.5.0.tar.gz"
 LIBDE265_URL = "https://github.com/strukturag/libde265/releases/download/v1.0.9/libde265-1.0.9.tar.gz"
-LIBHEIF_URL = "https://github.com/strukturag/libheif/archive/7caa01dd150b6c96f33d35bff2eab8a32b8edf2b.tar.gz"
+LIBHEIF_URL = "https://github.com/strukturag/libheif/archive/03fb8f092248bd4f76fad5399f4e3a9d3db4dba2.tar.gz"
 
 
 def download_file(url: str, out_path: str) -> bool: